Rename launcher and core modules
