+++ gui/dialogs/__init__.py
@@ -11,7 +11,7 @@ _module: ModuleType | None = None
 def __getattr__(name: str):
     global _module
     if _module is None:
-        _module = importlib.import_module("mainappsrc.AutoML")
+        _module = importlib.import_module("mainappsrc.automl_core")
     return getattr(_module, name)
