+++ mainappsrc/__init__.py
@@ -1,6 +1,6 @@
 """Main application package initializer."""
 
-from .AutoML import AutoMLApp
+from .automl_core import AutoMLApp
 from .page_diagram import PageDiagram
 from .fmeda_manager import FMEDAManager
 from .diagram_renderer import DiagramRenderer
